Update leave limit of carried forward leave rules from integer to double

+++ database/migrations/2021_09_28_182606_create_carried_forward_leave_rules_table.php
@@ -16,7 +16,7 @@ class CreateCarriedForwardLeaveRulesTable extends Migration
     {
         Schema::create('carried_forward_leave_rule', function (Blueprint $table) {
             $table->id();
-            $table->integer('leaveLimit');
+            $table->double('leaveLimit');
             $table->date('useBefore');
             $table->integer('approval');
             $table->date('startDate');
